requirements.txt
Django==5.0.4
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.0
django-cors-headers==4.3.1
firebase-admin==6.2.0